python-services/requirements.txt
fastapi
uvicorn
spacy
faiss-cpu
sentence-transformers 